refactor: MemberRegisterRequest 패키지 이동
