+++ apps/commerce-api/src/main/java/com/loopers/application/MemberFacade.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import com.loopers.application.provided.MemberFinder;
 import com.loopers.application.provided.MemberRegister;
 import com.loopers.domain.member.Member;
-import com.loopers.interfaces.api.member.dto.MemberRegisterRequest;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto.Request.MemberRegisterRequest;
 
 import lombok.RequiredArgsConstructor;
